docs :: move code
move to common dir
